Age Limit
Chef wants to appear in a competitive exam. To take the exam, there are the following requirements:

The minimum age limit is XX (i.e. Age should be greater than or equal to XX).
Age should be strictly less than YY.
The chef's current Age is AA. Find whether he is currently eligible to take the exam or not