DARK LIGHT
Tonmoy has a special torch. The torch has 44 levels numbered 11 to 44 and 22 states (\texttt{On}On and \texttt{Off}Off). Levels 1, 2,1,2, and 33 correspond to the \texttt{On}On state while level 44 corresponds to the \texttt{Off}Off state.

The levels of the torch can be changed as:

Level 11 changes to level 22.
Level 22 changes to level 33.
Level 33 changes to level 44.
Level 44 changes to level 11.
Given the initial state as KK and the number of changes made in the levels as NN, find the final state of the torch. If the final state cannot be determined, print \texttt{Ambiguous}Ambiguous instead.