CSTOCK
Chef wants to buy a stock whose price was SS rupees when the market opened. He will buy the stock if and only if its price is in the range [A, B][A,B]. The price of the stock has changed by C\%C% by the time he was trying to buy the stock. Will he be able to buy the stock?