KITCHENCOST
The chef visited a grocery store for fresh supplies. There are N items in the store where the I(the) item has a freshness value A(i) and cost B(i).
Chef has decided to purchase all the items having a freshness value greater than equal to X. Find the total cost of the groceries Chef buys.