NFS
The chef is playing Need For Speed. Currently, his car is running on a straight road with a velocity of U meters per second and approaching a 90∘ turn which is S meters away from him. To successfully cross the turn, the car's velocity when entering the turn must not exceed V meters per second.
The brakes of Chef's car allow him to slow down with a deceleration (negative acceleration) not exceeding A meters per squared second. Could you tell him whether he can cross the turn successfully? The velocity v when entering the arch can be determined from Newton's 2nd law to be v^2 = U^2 + 2⋅a⋅S if the car is moving with a uniform acceleration a.